system.prop: disable thread throttling on audio
This is probably why the audio stutters. AudioFlinger doesn't wanna
fill the stuff with data so quickly so it just waits a little bit
instead of continuing.

This property disables that behavior, allowing us to prevent audio
stutters and having a smooth playback. Although this reduces the
stutter, it does not fully fix it if other things are causing it
as well. This is just to at least prevent AudioFlinger from breaking
things more than they already are.

Change-Id: I7cae9bbd25634b50f7fa290fa75b655b47bd09c5